Add warning if one of the directories to backup doesn't exist

+++ scripts/deploy/backup-app.sh
@@ -6,5 +6,9 @@ mkdir -p ${webdir}_old && rm -rf ${webdir}_old/*
 
 for f in app assets scripts test index.php composer.json docker-compose.yml deployment_commit_id.txt
 do
-  cp -rp "${webdir}/$f" ${webdir}_old
+  if [ -d "${webdir}/$f" ] || [ -f "${webdir}/$f" ]; then
+    cp -rp "${webdir}/$f" ${webdir}_old
+  else
+    echo "Warning: ${webdir}/$f does not exist"
+  fi
 done
